Refactoring of python parts is ready for test Scripting and install stuff remains

+++ dbackup/location/localLocation.py
@@ -5,20 +5,21 @@ import shutil
 
 class LocalLocation(Location):
 
-    def __init__(self, spec, dynamichost = None, sshArgs = None):
+    def __init__(self, spec, dynamichost = None, sshArgs = None, simulate = False):
         """
-
         Arguments:
             spec (str) :  path to location
 
         """
         #self.type = 'remote' if re.match(r'^[^@:]*@[^@:]*:.*$', spec) is not None else 'local'
 
-        super().__init__(spec, dynamichost, sshArgs)
+        super().__init__(spec, dynamichost, typeName='local', simulate=simulate)
 
         self.path = spec
 
-    #overload
+    def __str__(self):
+        return f'LocalLocation:{self.path}'
+
     def rsyncPath(self, subpath = None):
         return self.path if subpath is None else os.path.join(self.path, subpath)
 
@@ -34,14 +35,15 @@ class LocalLocation(Location):
         else:
             logging.info("Creating local folder "+self.path)
             try:
-                os.makedirs(self.path)
+                if not self.simulate:
+                    os.makedirs(self.path)
                 return True
             except:
                 logging.error('ERROR: Could not create directory '+self.path)
             return False
 
 
-    def listDirs(self):
+    def listDir(self):
         """ Lists the directories in the location 
 
         Returns a list of the subdirectory names
@@ -57,7 +59,6 @@ class LocalLocation(Location):
 
     def renameChild(self, fromName, toName):
         """ Renames a file/folder in location
-
         """
         
         assert fromName != toName
@@ -68,5 +69,23 @@ class LocalLocation(Location):
         if os.path.exists(toPath):
             # Remove old instance if it already exists
             logging.debug('Replacing backup %s', toPath)
-            shutil.rmtree(toPath)
-        os.rename(fromPath, toPath)
+            if not self.simulate:
+                shutil.rmtree(toPath)
+        if not self.simulate:
+            os.rename(fromPath, toPath)
+
+    def deleteChild(self, name):
+        """ Removes a file/folder in the location """
+
+        if isinstance(name, str):
+            name = [name]
+        try:
+            for n in name:
+                filePath = os.path.join(self.path, n)
+                logging.debug('Removing ' + filePath)
+                if not self.simulate:
+                    shutil.rmtree(filePath)
+            return True
+        except PermissionError as e:
+            logging.error('Permission denied: %s', e.filename)
+        return False
